Move demo data loading to DemoDataLoader
